Update test classes
